modules/image/detection: fix function casts
The return value of a function should only be casted to a compatible
type; lots of "(int) floor(x)" can be simply changed to the implicit
casting rules.

+++ modules/src/image/detection/falign.c
@@ -141,8 +141,8 @@ Flist falign(Fimage u, int *d, int *nl, double *eps, float *g, char *all, Flists
   nx = u->ncol;
   ny = u->nrow;
   /* maximal length for a line */
-  n = (int) ceil(sqrt((double) nx * (double) nx 
-		      + (double) ny * (double) ny)) + 1;
+  n = ceil(sqrt((double) nx * (double) nx 
+		+ (double) ny * (double) ny)) + 1;
 
   /*** compute angle map of u ***/
   angle = ll_angle(u,*g);
